Mark 4 tests as XFAIL.
The three MSA tests are fragile and seem to have been changed by something in SelectionDAG (though the output still appears valid).

The remaining test, for the emergency spill slot, is failing because we handle emergency spill slots differently in CHERI-LLVM due to the small immediate size for capability instructions.